requirements.txt
django
djangorestframework
flake8
black
selenium
coverage
psycopg2-binary
